+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.2.0
+ * Refactoring to es6 class syntax, dependencies updated
+ * Add 'reversed' property.
+ * Add 'returnScrollable' method.
+
 0.1.2
  * fix issue when element is removed from list #14
  * Favor passive onScroll listener to polling when available. #15